# api-spec.yaml
openapi: 3.0.0
info:
  title: PeerPrep API
  description: API specification for PeerPrep G24 AY25/26
  version: 1.0.0
servers:
  - url: http://localhost:8001
    description: User Service
  - url: http://localhost:3002
    description: Matching Service
  - url: http://localhost:8000
    description: Question Service
  - url: http://localhost:3010
    description: Code Execution Service
  - url: http://localhost:8005
    description: Collaboration Service (future)
  - url: http://localhost:8006
    description: Queue Service (Internal, future)
...
paths:
...
  /api/questions:
    get:
      tags:
        - Questions
      summary: Get all questions (public)
      description: Retrieve all questions with optional filters for search, difficulty, topics, and pagination
      parameters:
        - name: search
          in: query
          description: Search by question title or topic keywords
          schema:
            type: string
        - name: difficulty
          in: query
          description: Filter by difficulty level
          schema:
            type: string
            enum: [EASY, MEDIUM, HARD]
        - name: topics
          in: query
          description: Filter by one or more topics
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of questions per page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestionSummary"
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /api/questions/{qid}:
    get:
      tags:
        - Questions
      summary: Get question by ID
      description: Retrieve detailed information about a specific question including description, examples, constraints, and test cases
      parameters:
        - name: qid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Question not found
...
  /v1/admin/questions:
    get:
      tags:
        - Questions - Admin
      summary: Get all questions with admin details (Admin)
      description: Retrieve all questions including soft-deleted ones and additional metadata
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Search by question title or topic keywords
          schema:
            type: string
        - name: difficulty
          in: query
          description: Filter by difficulty level
          schema:
            type: string
            enum: [EASY, MEDIUM, HARD]
        - name: topics
          in: query
          description: Filter by one or more topics
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: includeDeleted
          in: query
          description: Include soft-deleted questions
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of questions per page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Questions list with admin metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminQuestionSummary"
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

    post:
      tags:
        - Questions - Admin
      summary: Create new question (Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "400":
          description: Invalid input
  /v1/admin/questions/{qid}:
    get:
      tags:
        - Questions - Admin
      summary: Get question by ID with admin details (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: qid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question details with admin metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminQuestion"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Question not found

    put:
      tags:
        - Questions - Admin
      summary: Update question (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: qid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Question updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Question not found

    delete:
      tags:
        - Questions - Admin
      summary: Delete question (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: qid
          in: path
          required: true
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
            enum: [soft, hard]
            default: soft
          description: Soft delete (marks as deleted) or hard delete (permanently removes)
      responses:
        "204":
          description: Question deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Question not found
...
  /v1/admin/topics:
    get:
      tags:
        - Questions - Admin
      summary: Get all topics with usage stats (Admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Topics list with usage stats
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    questionCount:
                      type: integer
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

    post:
      tags:
        - Questions - Admin
      summary: Create new topic (Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Topic created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /v1/admin/topics/{topicName}:
    delete:
      tags:
        - Questions - Admin
      summary: Delete topic (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: topicName
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Topic deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ========== USER SERVICE - ADMIN ENDPOINTS ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    QuestionSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        topics:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
    Question:
      allOf:
        - $ref: "#/components/schemas/QuestionSummary"
        - type: object
          properties:
            description:
              type: string
            examples:
              type: array
              items:
                type: object
                properties:
                  input:
                    type: string
                  output:
                    type: string
                  explanation:
                    type: string
            constraints:
              type: array
              items:
                type: string
            testCases:
              type: array
              items:
                type: object
                properties:
                  input:
                    type: string
                  expectedOutput:
                    type: string
    QuestionInput:
      type: object
      required:
        - title
        - difficulty
        - topics
        - description
      properties:
        title:
          type: string
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        topics:
          type: array
          items:
            type: string
        description:
          type: string
        examples:
          type: array
          items:
            type: object
            properties:
              input:
                type: string
              output:
                type: string
              explanation:
                type: string
        constraints:
          type: array
          items:
            type: string
        testCases:
          type: array
          items:
            type: object
            properties:
              input:
                type: string
              expectedOutput:
                type: string
...
    AdminQuestionSummary:
      allOf:
        - $ref: "#/components/schemas/QuestionSummary"
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            isDeleted:
              type: boolean
            totalTestCases:
              type: integer
    AdminQuestion:
      allOf:
        - $ref: "#/components/schemas/Question"
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            createdBy:
              type: string
              description: User ID of the admin who created this question
            isDeleted:
              type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ForbiddenError:
      description: Insufficient permissions (Admin access required)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"